Refactor Dashboard components: Extract GuestUserNote, HiddenDashboardItems, OrderableRowItem, and ResetOrderButton into separate files; update dashboard styles; improve code organization and readability.
