+++ src/app/(screens)/dashboard.tsx
@@ -1,4 +1,9 @@
-import Divider from '@/components/Universal/Divider'
+import {
+	GuestUserNote,
+	HiddenDashboardItems,
+	ResetOrderButton,
+	dashboardStyles
+} from '@/components/Dashboard'
 import PlatformIcon from '@/components/Universal/Icon'
 import type { Card, ExtendedCard } from '@/components/all-cards'
 import { DashboardContext, UserKindContext } from '@/components/contexts'
@@ -8,21 +13,22 @@ import { USER_GUEST } from '@/data/constants'
 import { arraysEqual } from '@/utils/app-utils'
 import { toast } from 'burnt'
 import * as Haptics from 'expo-haptics'
-import { router } from 'expo-router'
-import React, { useCallback, useContext, useEffect, useState } from 'react'
+import type React from 'react'
+import { useCallback, useContext, useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import {
 	Dimensions,
+	type GestureResponderEvent,
 	LayoutAnimation,
 	Platform,
-	Pressable,
 	Text,
+	TouchableOpacity,
 	View
 } from 'react-native'
 import { DragSortableView } from 'react-native-drag-sort'
 import { ScrollView } from 'react-native-gesture-handler'
 import { runOnJS, runOnUI, useSharedValue } from 'react-native-reanimated'
-import { createStyleSheet, useStyles } from 'react-native-unistyles'
+import { useStyles } from 'react-native-unistyles'
 
 const { width } = Dimensions.get('window')
 
@@ -38,7 +44,7 @@ export default function DashboardEdit(): React.JSX.Element {
 		updateDashboardOrder
 	} = useContext(DashboardContext)
 	const { userKind = USER_GUEST } = useContext(UserKindContext)
-	const { styles, theme } = useStyles(stylesheet)
+	const { styles, theme } = useStyles(dashboardStyles)
 	const { t } = useTranslation(['settings'])
 	const [draggedId, setDraggedId] = useState<number | null>(null)
 	const [hasUserDefaultOrder, setHasUserDefaultOrder] = useState(true)
@@ -115,7 +121,7 @@ export default function DashboardEdit(): React.JSX.Element {
 		if (Platform.OS === 'ios') {
 			void Haptics.notificationAsync(Haptics.NotificationFeedbackType.Success)
 		}
-	}, [resetOrder])
+	}, [resetOrder, userKind])
 
 	useEffect(() => {
 		const { hidden, shown } = getDefaultDashboardOrder(userKind)
@@ -162,38 +168,8 @@ export default function DashboardEdit(): React.JSX.Element {
 				scrollEnabled={draggedId === null}
 			>
 				<View style={styles.wrapper}>
-					{userKind === USER_GUEST && (
-						<Pressable
-							style={[styles.card, styles.noteContainer]}
-							onPress={() => {
-								router.navigate('/login')
-							}}
-						>
-							<View style={styles.noteTextContainer}>
-								<PlatformIcon
-									ios={{
-										name: 'lock',
-										size: 20
-									}}
-									android={{
-										name: 'lock',
-										size: 24
-									}}
-									web={{
-										name: 'Lock',
-										size: 24
-									}}
-								/>
-								<Text style={styles.notesTitle}>
-									{t('dashboard.unavailable.title')}
-								</Text>
-							</View>
+					{userKind === USER_GUEST && <GuestUserNote />}
 
-							<Text style={styles.notesMessage}>
-								{t('dashboard.unavailable.message')}
-							</Text>
-						</Pressable>
-					)}
 					<View style={styles.block}>
 						<Text style={styles.sectionHeaderText}>{t('dashboard.shown')}</Text>
 						<View style={[styles.card, styles.shownBg]}>
@@ -201,7 +177,6 @@ export default function DashboardEdit(): React.JSX.Element {
 								<View
 									style={{
 										height: childrenHeight * 1.5,
-
 										...styles.emptyContainer
 									}}
 								>
@@ -272,115 +247,16 @@ export default function DashboardEdit(): React.JSX.Element {
 						</View>
 					</View>
 
-					<View style={styles.block}>
-						{filteredHiddenDashboardEntries.filter(Boolean).length > 0 && (
-							<Text style={styles.sectionHeaderText}>
-								{t('dashboard.hidden')}
-							</Text>
-						)}
-						<View style={styles.card}>
-							{filteredHiddenDashboardEntries
-								.filter(Boolean)
-								.map((item, index) => {
-									return (
-										<React.Fragment key={index}>
-											<Pressable
-												disabled={!item.removable}
-												onPress={() => {
-													handleRestore(item)
-												}}
-												hitSlop={10}
-												style={({ pressed }) => [
-													{
-														opacity: pressed ? 0.5 : 1,
-														minHeight: 46,
-														justifyContent: 'center'
-													}
-												]}
-											>
-												<View style={styles.row}>
-													<PlatformIcon
-														style={styles.minusIcon}
-														ios={{
-															name: cardIcons[
-																item.key as keyof typeof cardIcons
-															].ios,
-															size: 17
-														}}
-														android={{
-															name: cardIcons[
-																item.key as keyof typeof cardIcons
-															].android,
-															size: 21,
-															variant: 'outlined'
-														}}
-														web={{
-															name: cardIcons[
-																item.key as keyof typeof cardIcons
-															].web,
-															size: 21
-														}}
-													/>
-													<Text style={styles.text}>
-														{t(
-															// @ts-expect-error cannot verify the type
-															`cards.titles.${item.key}`,
-															{ ns: 'navigation' }
-														)}
-													</Text>
-													{!item.removable ? (
-														<PlatformIcon
-															style={styles.minusIcon}
-															ios={{
-																name: 'lock',
-																size: 20
-															}}
-															android={{
-																name: 'lock',
-																size: 24
-															}}
-															web={{
-																name: 'Lock',
-																size: 24
-															}}
-														/>
-													) : (
-														<PlatformIcon
-															ios={{
-																name: 'plus.circle',
-																variant: 'fill',
-																size: 20
-															}}
-															android={{
-																name: 'add_circle',
-																size: 24
-															}}
-															web={{
-																name: 'CirclePlus',
-																size: 24
-															}}
-															style={styles.restoreIcon}
-														/>
-													)}
-												</View>
-											</Pressable>
-											{index !== filteredHiddenDashboardEntries.length - 1 && (
-												<Divider width={'100%'} />
-											)}
-										</React.Fragment>
-									)
-								})}
-						</View>
-					</View>
-					{!hasUserDefaultOrder && (
-						<View style={[styles.card, styles.blockContainer]}>
-							<Pressable onPress={handleReset} disabled={hasUserDefaultOrder}>
-								<Text style={styles.reset(hasUserDefaultOrder)}>
-									{t('dashboard.reset')}
-								</Text>
-							</Pressable>
-						</View>
-					)}
+					<HiddenDashboardItems
+						filteredHiddenDashboardEntries={filteredHiddenDashboardEntries}
+						handleRestore={handleRestore}
+					/>
+
+					<ResetOrderButton
+						hasUserDefaultOrder={hasUserDefaultOrder}
+						onPress={handleReset}
+					/>
+
 					<Text style={styles.footer}>{t('dashboard.footer')}</Text>
 				</View>
 			</ScrollView>
@@ -401,8 +277,15 @@ function RowItem({
 	isLast,
 	isDragged
 }: RowItemProps): React.JSX.Element {
-	const { styles, theme } = useStyles(stylesheet)
+	const { styles, theme } = useStyles(dashboardStyles)
 	const bottomWidth = isLast || isDragged ? 0 : 1
+
+	const handleDeletePress = (e: GestureResponderEvent) => {
+		e?.stopPropagation?.()
+		if (item.removable) {
+			onPressDelete()
+		}
+	}
 
 	return (
 		<View>
@@ -439,14 +322,12 @@ function RowItem({
 				/>
 
 				<Text style={styles.text}>{item.text}</Text>
-				<Pressable
-					onPress={onPressDelete}
+
+				<TouchableOpacity
+					onPress={handleDeletePress}
+					activeOpacity={0.5}
+					style={{ opacity: item.removable ? 1 : 0 }}
 					disabled={!item.removable}
-					style={({ pressed }) => [
-						{
-							opacity: pressed ? 0.5 : 1
-						}
-					]}
 					hitSlop={{
 						top: 13,
 						right: 15,
@@ -454,134 +335,24 @@ function RowItem({
 						left: 15
 					}}
 				>
-					{item.removable && (
-						<PlatformIcon
-							ios={{
-								name: 'minus.circle',
-								size: 20
-							}}
-							android={{
-								name: 'do_not_disturb_on',
-								variant: 'outlined',
-								size: 24
-							}}
-							web={{
-								name: 'CircleMinus',
-								size: 24
-							}}
-							style={styles.minusIcon}
-						/>
-					)}
-				</Pressable>
+					<PlatformIcon
+						ios={{
+							name: 'minus.circle',
+							size: 20
+						}}
+						android={{
+							name: 'do_not_disturb_on',
+							variant: 'outlined',
+							size: 24
+						}}
+						web={{
+							name: 'CircleMinus',
+							size: 24
+						}}
+						style={styles.minusIcon}
+					/>
+				</TouchableOpacity>
 			</View>
 		</View>
 	)
 }
-
-const stylesheet = createStyleSheet((theme) => ({
-	block: {
-		alignSelf: 'center',
-		gap: 6,
-		width: '100%'
-	},
-	blockContainer: {
-		backgroundColor: theme.colors.card,
-		marginTop: 6
-	},
-	card: {
-		borderRadius: theme.radius.md,
-		overflow: 'hidden',
-		paddingHorizontal: 0
-	},
-	emptyContainer: {
-		backgroundColor: theme.colors.card,
-		borderRadius: theme.radius.md,
-		justifyContent: 'center'
-	},
-	footer: {
-		color: theme.colors.labelColor,
-		fontSize: 12,
-		fontWeight: 'normal',
-		textAlign: 'left'
-	},
-	minusIcon: {
-		color: theme.colors.labelSecondaryColor
-	},
-	noteContainer: {
-		backgroundColor: theme.colors.card,
-		marginTop: 3,
-		paddingHorizontal: 12
-	},
-	noteTextContainer: {
-		alignItems: 'center',
-		flexDirection: 'row',
-		gap: 8,
-		justifyContent: 'flex-start',
-		paddingTop: 12,
-		paddingVertical: 9
-	},
-	notesMessage: {
-		color: theme.colors.text,
-		fontSize: 15,
-		marginBottom: 12,
-		textAlign: 'left'
-	},
-	notesTitle: {
-		color: theme.colors.primary,
-		fontSize: 17,
-		fontWeight: '600',
-		textAlign: 'left'
-	},
-	outer: {
-		borderRadius: theme.radius.md,
-		flex: 1,
-		overflow: 'hidden'
-	},
-	outerRow: {
-		borderColor: theme.colors.border
-	},
-	page: {
-		padding: theme.margins.page
-	},
-	reset: (hasUserDefaultOrder: boolean) => ({
-		fontSize: 16,
-		marginVertical: 13,
-		alignSelf: 'center',
-		color: hasUserDefaultOrder ? theme.colors.labelColor : theme.colors.text
-	}),
-	restoreIcon: {
-		color: theme.colors.text
-	},
-	row: {
-		alignItems: 'center',
-		backgroundColor: theme.colors.card,
-		flexDirection: 'row',
-		gap: 14,
-		justifyContent: 'center',
-		minHeight: 48,
-		paddingHorizontal: 16
-	},
-	sectionHeaderText: {
-		color: theme.colors.labelSecondaryColor,
-		fontSize: 13,
-		fontWeight: 'normal',
-		textTransform: 'uppercase'
-	},
-	shownBg: {
-		backgroundColor: theme.colors.background
-	},
-	text: {
-		color: theme.colors.text,
-		flexGrow: 1,
-		flexShrink: 1,
-		fontSize: 16
-	},
-	textEmpty: {
-		color: theme.colors.text,
-		fontSize: 16,
-		textAlign: 'center'
-	},
-	wrapper: {
-		gap: 14
-	}
-}))
