+++ src/app/(screens)/dashboard.web.tsx
@@ -1,27 +1,22 @@
+import {
+	GuestUserNote,
+	HiddenDashboardItems,
+	OrderableRowItem,
+	ResetOrderButton,
+	dashboardStyles
+} from '@/components/Dashboard'
 import Divider from '@/components/Universal/Divider'
-import PlatformIcon from '@/components/Universal/Icon'
 import type { Card, ExtendedCard } from '@/components/all-cards'
 import { DashboardContext, UserKindContext } from '@/components/contexts'
-import { cardIcons } from '@/components/icons'
 import { getDefaultDashboardOrder } from '@/contexts/dashboard'
 import { USER_GUEST } from '@/data/constants'
 import { arraysEqual } from '@/utils/app-utils'
 import * as Haptics from 'expo-haptics'
-import { router } from 'expo-router'
 import React, { useCallback, useContext, useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import {
-	Dimensions,
-	LayoutAnimation,
-	Platform,
-	Pressable,
-	Text,
-	View
-} from 'react-native'
+import { LayoutAnimation, Platform, Text, View } from 'react-native'
 import { ScrollView } from 'react-native-gesture-handler'
-import { createStyleSheet, useStyles } from 'react-native-unistyles'
-
-const { width } = Dimensions.get('window')
+import { useStyles } from 'react-native-unistyles'
 
 export default function DashboardEdit(): React.JSX.Element {
 	const childrenHeight = 48
@@ -35,7 +30,7 @@ export default function DashboardEdit(): React.JSX.Element {
 		updateDashboardOrder
 	} = useContext(DashboardContext)
 	const { userKind = USER_GUEST } = useContext(UserKindContext)
-	const { styles } = useStyles(stylesheet)
+	const { styles } = useStyles(dashboardStyles)
 	const { t } = useTranslation(['settings'])
 	const [hasUserDefaultOrder, setHasUserDefaultOrder] = useState(true)
 	const [unavailableCards, setUnavailableCards] = useState<Card[]>([])
@@ -165,38 +160,8 @@ export default function DashboardEdit(): React.JSX.Element {
 				contentInsetAdjustmentBehavior="automatic"
 			>
 				<View style={styles.wrapper}>
-					{userKind === USER_GUEST && (
-						<Pressable
-							style={[styles.card, styles.noteContainer]}
-							onPress={() => {
-								router.navigate('/login')
-							}}
-						>
-							<View style={styles.noteTextContainer}>
-								<PlatformIcon
-									ios={{
-										name: 'lock',
-										size: 20
-									}}
-									android={{
-										name: 'lock',
-										size: 24
-									}}
-									web={{
-										name: 'Lock',
-										size: 24
-									}}
-								/>
-								<Text style={styles.notesTitle}>
-									{t('dashboard.unavailable.title')}
-								</Text>
-							</View>
+					{userKind === USER_GUEST && <GuestUserNote />}
 
-							<Text style={styles.notesMessage}>
-								{t('dashboard.unavailable.message')}
-							</Text>
-						</Pressable>
-					)}
 					<View style={styles.block}>
 						<Text style={styles.sectionHeaderText}>{t('dashboard.shown')}</Text>
 						<View style={[styles.card, styles.shownBg]}>
@@ -240,412 +205,19 @@ export default function DashboardEdit(): React.JSX.Element {
 						</View>
 					</View>
 
-					<View style={styles.block}>
-						{filteredHiddenDashboardEntries.filter(Boolean).length > 0 && (
-							<Text style={styles.sectionHeaderText}>
-								{t('dashboard.hidden')}
-							</Text>
-						)}
-						<View style={styles.card}>
-							{filteredHiddenDashboardEntries
-								.filter(Boolean)
-								.map((item, index) => {
-									return (
-										<React.Fragment key={index}>
-											<Pressable
-												disabled={!item.removable}
-												onPress={() => {
-													handleRestore(item)
-												}}
-												hitSlop={10}
-												style={({ pressed }) => [
-													{
-														opacity: pressed ? 0.5 : 1,
-														minHeight: 46,
-														justifyContent: 'center'
-													}
-												]}
-											>
-												<View style={styles.row}>
-													<PlatformIcon
-														style={styles.minusIcon}
-														ios={{
-															name: cardIcons[
-																item.key as keyof typeof cardIcons
-															].ios,
-															size: 17
-														}}
-														android={{
-															name: cardIcons[
-																item.key as keyof typeof cardIcons
-															].android,
-															size: 21,
-															variant: 'outlined'
-														}}
-														web={{
-															name: cardIcons[
-																item.key as keyof typeof cardIcons
-															].web,
-															size: 21
-														}}
-													/>
-													<Text style={styles.text}>
-														{t(
-															// @ts-expect-error cannot verify the type
-															`cards.titles.${item.key}`,
-															{ ns: 'navigation' }
-														)}
-													</Text>
-													{!item.removable ? (
-														<PlatformIcon
-															style={styles.minusIcon}
-															ios={{
-																name: 'lock',
-																size: 20
-															}}
-															android={{
-																name: 'lock',
-																size: 24
-															}}
-															web={{
-																name: 'Lock',
-																size: 24
-															}}
-														/>
-													) : (
-														<PlatformIcon
-															ios={{
-																name: 'plus.circle',
-																variant: 'fill',
-																size: 20
-															}}
-															android={{
-																name: 'add_circle',
-																size: 24
-															}}
-															web={{
-																name: 'CirclePlus',
-																size: 24
-															}}
-															style={styles.restoreIcon}
-														/>
-													)}
-												</View>
-											</Pressable>
-											{index !== filteredHiddenDashboardEntries.length - 1 && (
-												<Divider width={'100%'} />
-											)}
-										</React.Fragment>
-									)
-								})}
-						</View>
-					</View>
-					{!hasUserDefaultOrder && (
-						<View style={[styles.card, styles.blockContainer]}>
-							<Pressable onPress={handleReset} disabled={hasUserDefaultOrder}>
-								<Text style={styles.reset(hasUserDefaultOrder)}>
-									{t('dashboard.reset')}
-								</Text>
-							</Pressable>
-						</View>
-					)}
+					<HiddenDashboardItems
+						filteredHiddenDashboardEntries={filteredHiddenDashboardEntries}
+						handleRestore={handleRestore}
+					/>
+
+					<ResetOrderButton
+						hasUserDefaultOrder={hasUserDefaultOrder}
+						onPress={handleReset}
+					/>
+
 					<Text style={styles.footer}>{t('dashboard.footer')}</Text>
 				</View>
 			</ScrollView>
 		</View>
 	)
 }
-
-interface OrderableRowItemProps {
-	item: ExtendedCard
-	index: number
-	isLast: boolean
-	onMoveUp: () => void
-	onMoveDown: () => void
-	onPressDelete: () => void
-	isFirstItem: boolean
-	isLastItem: boolean
-}
-
-function OrderableRowItem({
-	item,
-	isLast,
-	onMoveUp,
-	onMoveDown,
-	onPressDelete,
-	isFirstItem,
-	isLastItem
-}: OrderableRowItemProps): React.JSX.Element {
-	const { styles, theme } = useStyles(stylesheet)
-	const bottomWidth = isLast ? 0 : 1
-
-	return (
-		<View>
-			<View
-				style={[
-					styles.row,
-					styles.outerRow,
-					{
-						width: width - theme.margins.page * 2,
-						borderBottomWidth: bottomWidth
-					}
-				]}
-			>
-				<View style={styles.iconContainer}>
-					<PlatformIcon
-						ios={{
-							name: cardIcons[item.key as keyof typeof cardIcons].ios,
-							size: 17
-						}}
-						android={{
-							name: cardIcons[item.key as keyof typeof cardIcons].android,
-							size: 21,
-							variant: 'outlined'
-						}}
-						web={{
-							name: cardIcons[item.key as keyof typeof cardIcons].web,
-							size: 21
-						}}
-					/>
-				</View>
-
-				<Text style={styles.text}>{item.text}</Text>
-
-				<View style={styles.actionButtons}>
-					{/* Move Up Button */}
-					<Pressable
-						onPress={onMoveUp}
-						disabled={isFirstItem}
-						style={({ pressed }) => [
-							styles.arrowButton,
-							{
-								opacity: isFirstItem ? 0.3 : pressed ? 0.7 : 1
-							}
-						]}
-						accessibilityLabel="Move up"
-					>
-						<PlatformIcon
-							ios={{
-								name: 'chevron.up',
-								size: 16
-							}}
-							android={{
-								name: 'keyboard_arrow_up',
-								size: 20
-							}}
-							web={{
-								name: 'ChevronUp',
-								size: 18
-							}}
-							style={styles.arrowIcon}
-						/>
-					</Pressable>
-
-					{/* Move Down Button */}
-					<Pressable
-						onPress={onMoveDown}
-						disabled={isLastItem}
-						style={({ pressed }) => [
-							styles.arrowButton,
-							{
-								opacity: isLastItem ? 0.3 : pressed ? 0.7 : 1
-							}
-						]}
-						accessibilityLabel="Move down"
-					>
-						<PlatformIcon
-							ios={{
-								name: 'chevron.down',
-								size: 16
-							}}
-							android={{
-								name: 'keyboard_arrow_down',
-								size: 20
-							}}
-							web={{
-								name: 'ChevronDown',
-								size: 18
-							}}
-							style={styles.arrowIcon}
-						/>
-					</Pressable>
-
-					{/* Delete Button or empty spacer for alignment */}
-					<View style={styles.deleteButtonContainer}>
-						{item.removable ? (
-							<Pressable
-								onPress={onPressDelete}
-								style={({ pressed }) => [
-									styles.deleteButton,
-									{
-										opacity: pressed ? 0.5 : 1
-									}
-								]}
-								hitSlop={{
-									top: 13,
-									right: 15,
-									bottom: 13,
-									left: 15
-								}}
-								accessibilityLabel="Remove item"
-							>
-								<PlatformIcon
-									ios={{
-										name: 'minus.circle',
-										size: 20
-									}}
-									android={{
-										name: 'do_not_disturb_on',
-										variant: 'outlined',
-										size: 24
-									}}
-									web={{
-										name: 'CircleMinus',
-										size: 24
-									}}
-									style={styles.minusIcon}
-								/>
-							</Pressable>
-						) : null}
-					</View>
-				</View>
-			</View>
-		</View>
-	)
-}
-
-const stylesheet = createStyleSheet((theme) => ({
-	block: {
-		alignSelf: 'center',
-		gap: 6,
-		width: '100%'
-	},
-	blockContainer: {
-		backgroundColor: theme.colors.card,
-		marginTop: 6
-	},
-	card: {
-		borderRadius: theme.radius.md,
-		overflow: 'hidden',
-		paddingHorizontal: 0
-	},
-	emptyContainer: {
-		backgroundColor: theme.colors.card,
-		borderRadius: theme.radius.md,
-		justifyContent: 'center'
-	},
-	footer: {
-		color: theme.colors.labelColor,
-		fontSize: 12,
-		fontWeight: 'normal',
-		textAlign: 'left'
-	},
-	minusIcon: {
-		color: theme.colors.labelSecondaryColor
-	},
-	noteContainer: {
-		backgroundColor: theme.colors.card,
-		marginTop: 3,
-		paddingHorizontal: 12
-	},
-	noteTextContainer: {
-		alignItems: 'center',
-		flexDirection: 'row',
-		gap: 8,
-		justifyContent: 'flex-start',
-		paddingTop: 12,
-		paddingVertical: 9
-	},
-	notesMessage: {
-		color: theme.colors.text,
-		fontSize: 15,
-		marginBottom: 12,
-		textAlign: 'left'
-	},
-	notesTitle: {
-		color: theme.colors.primary,
-		fontSize: 17,
-		fontWeight: '600',
-		textAlign: 'left'
-	},
-	outer: {
-		borderRadius: theme.radius.md,
-		flex: 1,
-		overflow: 'hidden'
-	},
-	outerRow: {
-		borderColor: theme.colors.border
-	},
-	page: {
-		padding: theme.margins.page
-	},
-	reset: (hasUserDefaultOrder: boolean) => ({
-		fontSize: 16,
-		marginVertical: 13,
-		alignSelf: 'center',
-		color: hasUserDefaultOrder ? theme.colors.labelColor : theme.colors.text
-	}),
-	restoreIcon: {
-		color: theme.colors.text
-	},
-	row: {
-		alignItems: 'center',
-		backgroundColor: theme.colors.card,
-		flexDirection: 'row',
-		gap: 14,
-		justifyContent: 'center',
-		minHeight: 48,
-		paddingHorizontal: 16
-	},
-	sectionHeaderText: {
-		color: theme.colors.labelSecondaryColor,
-		fontSize: 13,
-		fontWeight: 'normal',
-		textTransform: 'uppercase'
-	},
-	shownBg: {
-		backgroundColor: theme.colors.background
-	},
-	text: {
-		color: theme.colors.text,
-		flexGrow: 1,
-		flexShrink: 1,
-		fontSize: 16
-	},
-	textEmpty: {
-		color: theme.colors.text,
-		fontSize: 16,
-		textAlign: 'center'
-	},
-	wrapper: {
-		gap: 14
-	},
-	actionButtons: {
-		flexDirection: 'row',
-		alignItems: 'center'
-	},
-	arrowButton: {
-		padding: 8,
-		marginHorizontal: 2,
-		justifyContent: 'center',
-		alignItems: 'center'
-	},
-	arrowIcon: {
-		color: theme.colors.text
-	},
-	iconContainer: {
-		width: 24,
-		alignItems: 'center',
-		justifyContent: 'center'
-	},
-	deleteButtonContainer: {
-		width: 24,
-		marginLeft: 8,
-		alignItems: 'center',
-		justifyContent: 'center'
-	},
-	deleteButton: {
-		justifyContent: 'center',
-		alignItems: 'center'
-	}
-}))
